api/requirements.txt
fastapi==0.103.0  # Latest version as of October 2023
uvicorn==0.22.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.23.3
aiohttp==3.8.5
pydantic==2.4.2  # Upgraded to latest version
pydantic-settings==2.0.3  # Now compatible
transformers==4.32.1
torch==2.0.1
pillow==10.0.0
supabase==1.0.3
elevenlabs==0.2.21
python-jose==3.3.0
loguru==0.7.0
pydub==0.25.1
uuid==1.30
psutil==5.9.5  # For admin stats